Complete peer-to-peer messaging (unidirectional)
Currently, it renegotiates the handshake with every message.  This will
not do.  Replies should be implemented.  Alice should cease sending
OHELLO prefix messages once she has received a message from Bob.  Bob
should not do bprepare/bfinish more than once per session.

+++ src/util.c
@@ -91,8 +91,8 @@ store16_le(uint8_t out[2], uint16_t in)
 uint16_t
 load16_le(const uint8_t s[2])
 {
-    return (uint32_t)s[0]
-        | ((uint32_t)s[1] <<  8);
+    return (uint16_t)s[0]
+        | ((uint16_t)s[1] <<  8);
 }
 
 void
